[COZY-483] feat: 리사이클러뷰 수정
